move events under methods
