+++ hydroflows/methods/coastal/coastal_design_events.py
@@ -11,8 +11,8 @@ from hydromt.stats import eva, get_peak_hydrographs, get_peaks
 from pydantic import model_validator
 
 from hydroflows._typing import FileDirPath, ListOfInt, ListOfStr, OutputDirPath
-from hydroflows.events import Event, EventSet
 from hydroflows.methods.coastal.coastal_utils import plot_hydrographs
+from hydroflows.methods.events import Event, EventSet
 from hydroflows.workflow.method import ExpandMethod
 from hydroflows.workflow.method_parameters import Parameters
