+++ hydroflows/methods/discharge/fluvial_design_events.py
@@ -13,7 +13,7 @@ from hydromt.stats import design_events, extremes, get_peaks
 from pydantic import PositiveInt, model_validator
 
 from hydroflows._typing import FileDirPath, ListOfInt, ListOfStr, OutputDirPath
-from hydroflows.events import Event, EventSet
+from hydroflows.methods.events import Event, EventSet
 from hydroflows.workflow.method import ExpandMethod
 from hydroflows.workflow.method_parameters import Parameters
