+++ hydroflows/methods/fiat/fiat_update.py
@@ -6,7 +6,7 @@ from typing import List, Literal, Optional, Union
 from hydromt_fiat.fiat import FiatModel
 
 from hydroflows._typing import FileDirPath, ListOfPath, OutputDirPath, WildcardPath
-from hydroflows.events import EventSet
+from hydroflows.methods.events import EventSet
 from hydroflows.methods.fiat.fiat_utils import copy_fiat_model
 from hydroflows.utils.path_utils import make_relative_paths
 from hydroflows.workflow.method import ReduceMethod
