+++ hydroflows/methods/sfincs/sfincs_update_forcing.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import Optional
 
 from hydroflows._typing import FileDirPath, JsonDict, OutputDirPath
-from hydroflows.events import Event
+from hydroflows.methods.events import Event
 from hydroflows.methods.sfincs.sfincs_utils import parse_event_sfincs
 from hydroflows.workflow.method import Method
 from hydroflows.workflow.method_parameters import Parameters
